// src/enums/global-permissions.enum.ts
export enum GlobalPermissions {
  All_Permissions = "administer all",
  View_Administration_Page = "view administration",
  View_System_Logs = "view system logs",
  Administer_Settings = "administer settings",
  Administer_APIKeys = "administer api keys",
  Administer_Stations = "administer stations",
  Administer_CustomFields = "administer custom fields",
  Administer_Backups = "administer backups",
  Administer_Storage_Locations = "administer storage locations",
}